+++ src/usb.rs
@@ -20,7 +20,6 @@ pub struct Peripheral {
     pub usb: USB,
     pub pin_dm: PA11<Input<Floating>>,
     pub pin_dp: PA12<Input<Floating>>,
-    pub remap: bool,
 }
 
 unsafe impl Sync for Peripheral {}
@@ -32,7 +31,7 @@ unsafe impl UsbPeripheral for Peripheral {
     const EP_MEMORY_SIZE: usize = 1024;
 
     fn enable() {
-        let rcc = unsafe { (&*RCC::ptr()) };
+        let rcc = unsafe { &*RCC::ptr() };
 
         cortex_m::interrupt::free(|_| {
             // Enable USB peripheral
@@ -41,16 +40,6 @@ unsafe impl UsbPeripheral for Peripheral {
             // Reset USB peripheral
             rcc.apb1rstr.modify(|_, w| w.usbrst().set_bit());
             rcc.apb1rstr.modify(|_, w| w.usbrst().clear_bit());
-
-            // Are we remapping USB pins?
-            if self.usb_remap {
-                // Remap PA11/PA12 pins to PA09/PA10 for USB on
-                // TSSOP20 (STM32F042F) or UFQFPN28 (STM32F042G) packages
-                let syscfg = unsafe { (&*SYSCFG::ptr()) };
-
-                rcc.apb2enr.modify(|_, w| w.syscfgen().set_bit());
-                syscfg.cfgr1.modify(|_, w| w.pa11_pa12_rmp().remapped());
-            }
         });
     }
 
@@ -59,6 +48,15 @@ unsafe impl UsbPeripheral for Peripheral {
         // at least that long.
         cortex_m::asm::delay(72);
     }
+}
+
+pub fn remap_pins(rcc: &mut RCC, syscfg: &mut SYSCFG) {
+    cortex_m::interrupt::free(|_| {
+        // Remap PA11/PA12 pins to PA09/PA10 for USB on
+        // TSSOP20 (STM32F042F) or UFQFPN28 (STM32F042G) packages
+        rcc.apb2enr.modify(|_, w| w.syscfgen().set_bit());
+        syscfg.cfgr1.modify(|_, w| w.pa11_pa12_rmp().remapped());
+    });
 }
 
 pub type UsbBusType = UsbBus<Peripheral>;