+++ examples/usb_serial.rs
@@ -15,13 +15,6 @@ use usbd_serial::{SerialPort, USB_CLASS_CDC};
 fn main() -> ! {
     let mut dp = pac::Peripherals::take().unwrap();
 
-    /* Uncomment the following lines if you have a chip in TSSOP20 (STM32F042F)
-       or UFQFPN28 (STM32F042G) package
-       This code enables clock for SYSCFG and remaps USB pins to PA9 and PA10.
-    */
-    //dp.RCC.apb2enr.modify(|_, w| w.syscfgen().set_bit());
-    //dp.SYSCFG.cfgr1.modify(|_, w| w.pa11_pa12_rmp().remapped());
-
     let mut rcc = dp
         .RCC
         .configure()
@@ -43,9 +36,19 @@ fn main() -> ! {
         pin_dm: gpioa.pa11,
         pin_dp: gpioa.pa12,
     };
+
     let usb_bus = UsbBus::new(usb);
 
     let mut serial = SerialPort::new(&usb_bus);
+
+    /*
+     * IMPORTANT: if you have a chip in TSSOP20 (STM32F042F) or UFQFPN28 (STM32F042G) package,
+     * and want to use USB, make sure you call `remap_pins(rcc, syscfg)`, otherwise the device will not enumerate.
+     *
+     * Uncomment the following function if the situation above applies to you.
+     */
+
+    //usb_bus.remap_pins();
 
     let mut usb_dev = UsbDeviceBuilder::new(&usb_bus, UsbVidPid(0x16c0, 0x27dd))
         .manufacturer("Fake company")